# openapi/web-v1.json
{
  "components": {
    "schemas": {
...
      "AdminStatusResponse": {
        "properties": {
          "database": {
            "title": "Database",
            "type": "boolean"
          },
          "embedding_model": {
            "title": "Embedding Model",
            "type": "string"
          },
          "embeddings": {
            "title": "Embeddings",
            "type": "boolean"
          },
          "model_mismatch": {
            "title": "Model Mismatch",
            "type": "boolean"
          }
        },
        "required": [
          "database",
          "embeddings",
          "embedding_model",
          "model_mismatch"
        ],
        "title": "AdminStatusResponse",
        "type": "object"
      },
      "AdminUpdate": {
        "properties": {
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          }
        },
        "required": [
          "is_admin"
        ],
        "title": "AdminUpdate",
        "type": "object"
      },
      "AggregatesResponse": {
        "properties": {
          "active_keys": {
            "title": "Active Keys",
            "type": "integer"
          },
          "memories": {
            "items": {
              "$ref": "#/components/schemas/MemoryVolume"
            },
            "title": "Memories",
            "type": "array"
          },
          "revoked_keys": {
            "title": "Revoked Keys",
            "type": "integer"
          },
          "total_users": {
            "title": "Total Users",
            "type": "integer"
          }
        },
        "required": [
          "total_users",
          "active_keys",
          "revoked_keys",
          "memories"
        ],
        "title": "AggregatesResponse",
        "type": "object"
      },
...
      "IssuedKeyResponse": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "last_used_at",
          "revoked",
          "api_key"
        ],
        "title": "IssuedKeyResponse",
        "type": "object"
      },
      "KeyIssue": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "KeyIssue",
        "type": "object"
      },
      "KeyResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "last_used_at",
          "revoked"
        ],
        "title": "KeyResponse",
        "type": "object"
      },
...
      "MemoryVolume": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "count"
        ],
        "title": "MemoryVolume",
        "type": "object"
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "is_admin": {
            "default": false,
            "title": "Is Admin",
            "type": "boolean"
          }
        },
        "required": [
          "email"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "active_key_count": {
            "title": "Active Key Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "web_access": {
            "title": "Web Access",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "created_at",
          "is_admin",
          "web_access",
          "active_key_count"
        ],
        "title": "UserResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "recallum_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Recallum Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/aggregates": {
      "get": {
        "operationId": "aggregates_admin_aggregates_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "ge": 1,
              "le": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregatesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Aggregates",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/status": {
      "get": {
        "operationId": "detailed_status_admin_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Detailed Status",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "list_users_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "ge": 1,
              "le": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response List Users Admin Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_user_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/admin": {
      "put": {
        "operationId": "set_admin_admin_users__user_id__admin_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Set Admin",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/keys": {
      "get": {
        "operationId": "list_keys_admin_users__user_id__keys_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KeyResponse"
                  },
                  "title": "Response List Keys Admin Users  User Id  Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Keys",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "issue_key_admin_users__user_id__keys_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Issue Key",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/keys/{key_id}/revoke": {
      "post": {
        "operationId": "revoke_key_admin_users__user_id__keys__key_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Revoke Key",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}